moved conditionaleventdefinition
