+++ tests/ProcessMaker/Models/RootElementRepository.php
@@ -2,7 +2,7 @@
 
 namespace ProcessMaker\Models;
 
-use ProcessMaker\Nayra\Bpmn\ConditionalEventDefinition;
+use ProcessMaker\Nayra\Bpmn\Models\ConditionalEventDefinition;
 use ProcessMaker\Nayra\Bpmn\ErrorEventDefinition;
 use ProcessMaker\Nayra\Bpmn\Models\FormalExpression;
 use ProcessMaker\Nayra\Bpmn\Models\ItemDefinition;
